refactor transaction processing; add more tests
