+++ subgraphs/magicswapv2/tests/vault.test.ts
@@ -1,19 +1,10 @@
-import {
-  assert,
-  beforeEach,
-  clearStore,
-  describe,
-  newMockEvent,
-  test,
-} from "matchstick-as";
+import { assert, beforeEach, clearStore, describe, test } from "matchstick-as";
 
-import { Address, Bytes, ethereum } from "@graphprotocol/graph-ts";
+import { Address, Bytes } from "@graphprotocol/graph-ts";
 
-import { VaultCreated } from "../generated/NftVaultFactory/NftVaultFactory";
 import { VaultCollection } from "../generated/schema";
-import { Deposit } from "../generated/templates/NftVault/NftVault";
 import { handlePairCreated } from "../src/mappings/factory";
-import { handleTransfer } from "../src/mappings/pair";
+import { handleMint, handleTransfer } from "../src/mappings/pair";
 import { handleDeposit, handleVaultCreated } from "../src/mappings/vault";
 import {
   COLLECTION,
@@ -23,72 +14,16 @@ import {
   USER1,
   VAULT,
 } from "./helpers/constants";
+import {
+  createMintEvent,
+  createPairCreatedEvent,
+  createTransferEvent,
+} from "./helpers/pair";
 import { mockToken } from "./helpers/token";
-import { createPairCreatedEvent, createTransferEvent } from "./pair.test";
+import { createDepositEvent, createVaultCreatedEvent } from "./helpers/vault";
 
+mockToken(TOKEN0, "Token 0", "TK0");
 mockToken(VAULT);
-
-const createVaultCreatedEvent = (
-  vault: string,
-  name: string,
-  symbol: string,
-  collectionAddr: string,
-  collectionNftType: i32,
-  collectionTokenIds: i32[]
-): VaultCreated => {
-  const event = changetype<VaultCreated>(newMockEvent());
-  event.parameters = [
-    new ethereum.EventParam("name", ethereum.Value.fromString(name)),
-    new ethereum.EventParam("symbol", ethereum.Value.fromString(symbol)),
-    new ethereum.EventParam(
-      "vault",
-      ethereum.Value.fromAddress(Address.fromString(vault))
-    ),
-    new ethereum.EventParam("vaultId", ethereum.Value.fromI32(0)),
-    new ethereum.EventParam(
-      "collections",
-      ethereum.Value.fromTupleArray([
-        changetype<ethereum.Tuple>([
-          ethereum.Value.fromAddress(Address.fromString(collectionAddr)),
-          ethereum.Value.fromI32(collectionNftType),
-          ethereum.Value.fromBoolean(collectionTokenIds.length === 0),
-          ethereum.Value.fromI32Array(collectionTokenIds),
-        ]),
-      ])
-    ),
-    new ethereum.EventParam(
-      "creator",
-      ethereum.Value.fromAddress(Address.zero())
-    ),
-  ];
-  return event;
-};
-
-const createDepositEvent = (
-  hash: string,
-  vault: string,
-  to: string,
-  collection: string,
-  tokenId: i32,
-  amount: i32
-): Deposit => {
-  const event = changetype<Deposit>(newMockEvent());
-  event.address = Address.fromString(vault);
-  event.transaction.hash = Bytes.fromHexString(hash);
-  event.parameters = [
-    new ethereum.EventParam(
-      "to",
-      ethereum.Value.fromAddress(Address.fromString(to))
-    ),
-    new ethereum.EventParam(
-      "collection",
-      ethereum.Value.fromAddress(Address.fromString(collection))
-    ),
-    new ethereum.EventParam("tokenId", ethereum.Value.fromI32(tokenId)),
-    new ethereum.EventParam("amount", ethereum.Value.fromI32(amount)),
-  ];
-  return event;
-};
 
 describe("handleVaultCreated()", () => {
   beforeEach(() => {
@@ -178,6 +113,14 @@ describe("handleDeposit()", () => {
         Address.zero().toHexString(),
         USER1,
         "1000000000000000000"
+      )
+    );
+    handleMint(
+      createMintEvent(
+        PAIR,
+        USER1,
+        "1000000000000000000",
+        "500000000000000000000"
       )
     );
